backend/pyproject.toml
[project]
name = "bi-dashboard-backend"
version = "0.1.0"
description = "Conversational AI for Instant Business Intelligence Dashboards"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.2.0",
    "groq>=0.11.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
